Move invariant to invariant package
