+++ modules/bank/invariants/invariants.go
@@ -1,4 +1,4 @@
-package bank
+package invariants
 
 import (
 	"errors"
