+++ modules/distribution/invariants/invariants.go
@@ -1,15 +1,16 @@
-package distribution
+package invariants
 
 import (
 	"fmt"
 
 	sdk "github.com/irisnet/irishub/types"
+	distr "github.com/irisnet/irishub/modules/distribution"
 	"github.com/irisnet/irishub/modules/mock/simulation"
 )
 
 // AllInvariants runs all invariants of the distribution module
 // Currently: total supply, positive power
-func AllInvariants(d Keeper, sk StakeKeeper) simulation.Invariant {
+func AllInvariants(d distr.Keeper, sk distr.StakeKeeper) simulation.Invariant {
 	return func(ctx sdk.Context) error {
 		err := ValAccumInvariants(d, sk)(ctx)
 		if err != nil {
@@ -20,13 +21,13 @@ func AllInvariants(d Keeper, sk StakeKeeper) simulation.Invariant {
 }
 
 // ValAccumInvariants checks that the fee pool accum == sum all validators' accum
-func ValAccumInvariants(k Keeper, sk StakeKeeper) simulation.Invariant {
+func ValAccumInvariants(k distr.Keeper, sk distr.StakeKeeper) simulation.Invariant {
 
 	return func(ctx sdk.Context) error {
 		height := ctx.BlockHeight()
 
 		valAccum := sdk.ZeroDec()
-		k.IterateValidatorDistInfos(ctx, func(_ int64, vdi ValidatorDistInfo) bool {
+		k.IterateValidatorDistInfos(ctx, func(_ int64, vdi distr.ValidatorDistInfo) bool {
 			lastValPower := sk.GetLastValidatorPower(ctx, vdi.OperatorAddr)
 			valAccum = valAccum.Add(vdi.GetValAccum(height, sdk.NewDecFromInt(lastValPower)))
 			return false
